Update tournament description
Update tournament confirmation description to prevent mistakes from end-users

Makes the description for the tournament character conversion more explicit in what it does and what you cannot do after conversion as well as illustrating in block capitals that it cannot be undone. 

Aims to prevent end-users mistakenly believing this can be used on cRPGs main servers.