webui: shorten item tier tooltip description (#1129)
Updates the Item Tier tooltip description to make it shorter and more succinct in its explanation. The previous tooltip contained a TODO tag and an overlong explanation of the item value model.

In the revised tooltip, I've replaced the TODO tag and have provided a more concise explanation of the item value model. I also highlight that it is not a perfect indicator of balance for players.